Switch type of priority from 'double' to 'unsigned long long'.

+++ src/pqueue.h
@@ -36,7 +36,7 @@
 #define PQUEUE_H
 
 /** priority data type */
-typedef double pqueue_pri_t;
+typedef unsigned long long pqueue_pri_t;
 
 /** callback functions to get/set/compare the priority of an element */
 typedef pqueue_pri_t (*pqueue_get_pri_f)(void *a);
@@ -56,15 +56,15 @@ typedef void (*pqueue_print_entry_f)(FILE *out, void *a);
 /** the priority queue handle */
 typedef struct pqueue_t
 {
-    size_t size;
-    size_t avail;
-    size_t step;
-    pqueue_cmp_pri_f cmppri;
-    pqueue_get_pri_f getpri;
-    pqueue_set_pri_f setpri;
-    pqueue_get_pos_f getpos;
-    pqueue_set_pos_f setpos;
-    void **d;
+    size_t size;                /**< number of elements in this queue */
+    size_t avail;               /**< slots available in this queue */
+    size_t step;                /**< growth stepping setting */
+    pqueue_cmp_pri_f cmppri;    /**< callback to compare nodes */
+    pqueue_get_pri_f getpri;    /**< callback to get priority of a node */
+    pqueue_set_pri_f setpri;    /**< callback to set priority of a node */
+    pqueue_get_pos_f getpos;    /**< callback to get position of a node */
+    pqueue_set_pos_f setpos;    /**< callback to set position of a node */
+    void **d;                   /**< The actualy queue in binary heap form */
 } pqueue_t;
 
 
@@ -72,12 +72,16 @@ typedef struct pqueue_t
  * initialize the queue
  *
  * @param n the initial estimate of the number of queue items for which memory
- *          should be preallocated
- * @param pri the callback function to run to assign a score to a element
- * @param get the callback function to get the current element's position
- * @param set the callback function to set the current element's position
+ *     should be preallocated
+ * @param cmppri The callback function to run to compare two elements
+ *     This callback should return 0 for 'lower' and non-zero
+ *     for 'higher', or vice versa if reverse priority is desired
+ * @param setpri the callback function to run to assign a score to an element
+ * @param getpri the callback function to run to set a score to an element
+ * @param getpos the callback function to get the current element's position
+ * @param setpos the callback function to set the current element's position
  *
- * @Return the handle or NULL for insufficent memory
+ * @return the handle or NULL for insufficent memory
  */
 pqueue_t *
 pqueue_init(size_t n,
@@ -114,7 +118,7 @@ int pqueue_insert(pqueue_t *q, void *d);
 /**
  * move an existing entry to a different priority
  * @param q the queue
- * @param old the old priority
+ * @param new_pri the new priority
  * @param d the entry
  */
 void
@@ -125,8 +129,7 @@ pqueue_change_priority(pqueue_t *q,
 
 /**
  * pop the highest-ranking item from the queue.
- * @param p the queue
- * @param d where to copy the entry to
+ * @param q the queue
  * @return NULL on error, otherwise the entry
  */
 void *pqueue_pop(pqueue_t *q);
@@ -134,7 +137,7 @@ void *pqueue_pop(pqueue_t *q);
 
 /**
  * remove an item from the queue.
- * @param p the queue
+ * @param q the queue
  * @param d the entry
  * @return 0 on success
  */
@@ -144,7 +147,6 @@ int pqueue_remove(pqueue_t *q, void *d);
 /**
  * access highest-ranking item without removing it.
  * @param q the queue
- * @param d the entry
  * @return NULL on error, otherwise the entry
  */
 void *pqueue_peek(pqueue_t *q);
@@ -160,7 +162,7 @@ void *pqueue_peek(pqueue_t *q);
  */
 void
 pqueue_print(pqueue_t *q, 
-             FILE *out, 
+             FILE *out,
              pqueue_print_entry_f print);
